js/xpconnect/tests/mochitest/test_bug912322.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=912322
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 912322</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="application/javascript">

  var { AppConstants } = SpecialPowers.ChromeUtils.importESModule(
    "resource://gre/modules/AppConstants.sys.mjs"
  );
  // Test window.controllers.
  if (AppConstants.RELEASE_OR_BETA) {
    is(typeof window.controllers, 'object', "shimmed controllers should be available to content in beta and release");
  } else {
    is(typeof window.controllers, 'undefined', "controllers should not be available to content in Nightly");
  }
  is(typeof SpecialPowers.wrap(window).controllers, 'object', "controllers should be available over Xray");

  </script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=912322">Mozilla Bug 912322</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
</pre>
</body>
</html>
